python2 compatibility fix
In py2, `round` with one argument returns float, whereas in py3, `round` returns int.

+++ src/main.py
@@ -431,7 +431,7 @@ class MainWindow(Qt.QMainWindow):
             self.glWidget.setUniform('time', self.time_uniform)
         self.time.start()
         self.glWidget.tick()
-        self.setWindowTitle("{:d} fps".format(round(self.glWidget.getFps())))
+        self.setWindowTitle("{:d} fps".format(int(round(self.glWidget.getFps()))))
 
     def timer_reset(self):
         self.time_uniform = 0.0
